requirements.txt
argon2-cffi==19.2.0
Babel==2.8.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
crypto==1.4.1
cryptography==2.8
Flask==1.1.1
Flask-Admin==1.5.4
Flask-BabelEx==0.9.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Naked==0.1.31
passlib==1.7.2
pycparser==2.19
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
shellescape==3.4.1
six==1.13.0
speaklater==1.3
SQLAlchemy==1.3.12
urllib3==1.25.7
virtualenv==16.7.9
Werkzeug==0.16.0
WTForm==1.0
WTForms==2.2.1
